Functional
The PHP and MYSQLi work.  So you can signup, and then login and retreive that data.

There are still some bugs to fix, i.e. even if the passwords don't match or there's some other error picked up in one of the fields, the info will still post to the DB.  It could also use some more error messages and checks when connecting.